Fix warnings, update version
Fix a sprintf format warning without changing the actual functionality
Putting statics in headers is re-tard-ed (json_spirit_value)
This is alpha 3 of 1.5

+++ src/irc.cpp
@@ -255,7 +255,7 @@ void ThreadIRCSeed2()
         if (GetLocal(addrLocal, &addrIPv4))
             strMyName = EncodeAddress(GetLocalAddress(&addrConnect));
         if (strMyName == "")
-            strMyName = strprintf("x%u", GetRand(1000000000));
+            strMyName = strprintf("x%u", (uint32_t)GetRand(1000000000));
 
         Send(hSocket, strprintf("NICK %s\r", strMyName.c_str()).c_str());
         Send(hSocket, strprintf("USER %s 8 * : %s\r", strMyName.c_str(), strMyName.c_str()).c_str());
